1. add format currency pipe 2. add soft delete

+++ src/moma.be/Motel/Motel.Application/Services/ContractService/Dtos/ContractDto.cs
@@ -25,7 +25,7 @@ namespace Motel.Application.Services.ContractService.Dtos
 
         public DateTime? ExpiredDate { get; set; }
 
-        public double DepositedAmount { get; set; }
+        public decimal DepositedAmount { get; set; }
 
         public EnumContractType Type { get; set; }
 
@@ -35,7 +35,7 @@ namespace Motel.Application.Services.ContractService.Dtos
 
         public List<AddTermDto> Terms { get; set; }
 
-        public double AdvanceAmount { get; set; }
+        public decimal AdvanceAmount { get; set; }
 
         public static ContractDto FromDeposited(RoomDeposited deposited)
         {
