+++ src/moma.be/Motel/Motel.Application/Services/InvoiceService/Dtos/InvoiceRoomDto.cs
@@ -16,13 +16,13 @@ namespace Motel.Application.Services.InvoiceService.Dtos
 
         public Guid? ProvideId { get; set; }
 
-        public double? LastValue { get; set; }
+        public decimal? LastValue { get; set; }
 
-        public double? NewValue { get; set; }
+        public decimal? NewValue { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double? Amount { get; set; }
+        public decimal? Amount { get; set; }
 
         public string Name { get; set; }
