+++ src/moma.be/Motel/Motel.Application/Services/ProvideService/Dtos/ServiceDto.cs
@@ -15,7 +15,7 @@ namespace Motel.Application.Services.ServiceService.Dtos
 
         public EnumServiceType Type { get; set; }
 
-        public double DefaultPrice { get; set; }
+        public decimal DefaultPrice { get; set; }
     }
 
     public class ProvideModel
@@ -23,6 +23,6 @@ namespace Motel.Application.Services.ServiceService.Dtos
         public Guid? Id { get; set; }
         public string Name { get; set; }
         public EnumServiceType Type { get; set; }
-        public double DefaultPrice { get; set; }
+        public decimal DefaultPrice { get; set; }
     }
 }
