+++ src/moma.be/Motel/Motel.Application/Services/StageService/Dtos/StageDto.cs
@@ -11,9 +11,9 @@ namespace Motel.Application.Services.StageService.Dtos
 
         public string Name { get; set; }
 
-        public double TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
 
-        public double AmountPaid { get; set; }
+        public decimal AmountPaid { get; set; }
 
         public int RoomData { get; set; }
